Rewrite sorting algorithm
When working with 'char' based movements that assumption prior was that
the entire list was delimited by ', '. Issue #8 asked for handling
alternative delimiters. I think this algorithm can accomplish that.

It will also allow the use of the around motion as it preserves prefix
and suffix characters. The only assumption is that the selection is
delimited consistently If the first delimiter is ', ' then all of them
much match the same or things go wonky. (Future improvement maybe)

This also moves the cursor back to the beginning like most Vim motion
commands do.